refactor(scaffold): relocate out of engines/pi to a neutral home
The scaffold writes SPEC-level workspace shape (AGENTS.md, skills/, tools/, channels/), not
pi internals — it does not belong under engines/pi. Move src/engines/pi/scaffold/ →
src/scaffold/ and untangle the coupling that kept it there:

- assertInsideWorkspace + loadRootIgnore were engine-neutral (pure fs/path + the `ignore`
  matcher) but lived in engines/pi/definition.ts and were shared by the engine's channel
  discovery AND the scaffold. Extract them to a neutral src/workspace.ts.
- fastagentVersion (the package's own version, read for the scaffolded dependency range) was
  neutral but under engines/pi; move it to src/version.ts.
- scaffold now depends on the engine only through vendor-skill's loadAgentDefinition — the one
  legitimate edge, since vendoring a skill validates it loads under the engine that will run it.

Pure relocation: imports repointed (channel.ts, cli.ts, index.ts, test), definition.ts loses
the two functions and their now-unused imports. typecheck/lint/248 tests green; the built CLI
init + --version work from dist.
