+++ core/src/scaffold/scaffold-templates.ts
