+++ core/test/init.test.ts
@@ -5,8 +5,8 @@ import { tmpdir } from "node:os";
 import { join } from "node:path";
 import { fileURLToPath } from "node:url";
 import { createPiAgentFromWorkspace, loadAgentDefinition, scaffoldWorkspace } from "../src/index.ts";
-import { nextStepCd } from "../src/engines/pi/scaffold/init.ts";
-import { vendorSkill } from "../src/engines/pi/scaffold/vendor-skill.ts";
+import { nextStepCd } from "../src/scaffold/init.ts";
+import { vendorSkill } from "../src/scaffold/vendor-skill.ts";
 
 const freshDir = () => mkdtemp(join(tmpdir(), "fa-init-"));
 async function exists(p: string): Promise<boolean> {
